requirements.txt
tk
paho-mqtt
playsound
